+++ pylabrobot/agilent/biotek/__init__.py
@@ -1,9 +1,6 @@
 from .biotek import BioTekBackend
-from .cytation import (
-  Cytation1,
-  Cytation5,
-  CytationBackend,
-  CytationImagingConfig,
-)
+from .cytation import CytationBackend, CytationImagingConfig
+from .cytation1 import Cytation1
+from .cytation5 import Cytation5
 from .el406 import EL406, EL406Driver, EL406PlateWashingBackend, EL406ShakingBackend
 from .synergy_h1 import SynergyH1, SynergyH1Backend
