+++ pylabrobot/agilent/biotek/cytation.py
@@ -885,107 +885,8 @@ class CytationBackend(BioTekBackend, MicroscopyBackend):
 
 
 # ---------------------------------------------------------------------------
-# Devices
+# Backwards compatibility — device classes moved to cytation5.py / cytation1.py
 # ---------------------------------------------------------------------------
 
-
-class Cytation5(Resource, Device):
-  """Agilent BioTek Cytation 5 — plate reader + imager."""
-
-  def __init__(
-    self,
-    name: str,
-    device_id: Optional[str] = None,
-    imaging_config: Optional[CytationImagingConfig] = None,
-    size_x: float = 0.0,  # TODO: measure
-    size_y: float = 0.0,  # TODO: measure
-    size_z: float = 0.0,  # TODO: measure
-  ):
-    backend = CytationBackend(device_id=device_id, imaging_config=imaging_config)
-    Resource.__init__(
-      self,
-      name=name,
-      size_x=size_x,
-      size_y=size_y,
-      size_z=size_z,
-      model="Agilent BioTek Cytation 5",
-    )
-    Device.__init__(self, driver=backend)
-    self.driver: CytationBackend = backend
-    self.absorbance = Absorbance(backend=backend)
-    self.luminescence = Luminescence(backend=backend)
-    self.fluorescence = Fluorescence(backend=backend)
-    self.microscopy = Microscopy(backend=backend)
-    self.temperature = TemperatureController(backend=backend)
-    self._capabilities = [
-      self.absorbance,
-      self.luminescence,
-      self.fluorescence,
-      self.microscopy,
-      self.temperature,
-    ]
-
-    self.plate_holder = PlateHolder(
-      name=name + "_plate_holder",
-      size_x=127.76,
-      size_y=85.48,
-      size_z=0,  # TODO: measure
-      pedestal_size_z=0,
-      child_location=Coordinate.zero(),  # TODO: measure
-    )
-    self.assign_child_resource(self.plate_holder, location=Coordinate.zero())
-
-  def serialize(self) -> dict:
-    return {**Resource.serialize(self), **Device.serialize(self)}
-
-  async def open(self, slow: bool = False) -> None:
-    await self.driver.open(slow=slow)
-
-  async def close(self, slow: bool = False) -> None:
-    await self.driver.close(slow=slow)
-
-
-class Cytation1(Resource, Device):
-  """Agilent BioTek Cytation 1 — plate reader only (no imager)."""
-
-  def __init__(
-    self,
-    name: str,
-    device_id: Optional[str] = None,
-    size_x: float = 0.0,  # TODO: measure
-    size_y: float = 0.0,  # TODO: measure
-    size_z: float = 0.0,  # TODO: measure
-  ):
-    backend = BioTekBackend(device_id=device_id)
-    Resource.__init__(
-      self,
-      name=name,
-      size_x=size_x,
-      size_y=size_y,
-      size_z=size_z,
-      model="Agilent BioTek Cytation 1",
-    )
-    Device.__init__(self, driver=backend)
-    self.driver: BioTekBackend = backend
-    self.microscopy = Microscopy(backend=backend)  # type: ignore[arg-type]
-    self.temperature = TemperatureController(backend=backend)
-    self._capabilities = [self.microscopy, self.temperature]
-
-    self.plate_holder = PlateHolder(
-      name=name + "_plate_holder",
-      size_x=127.76,
-      size_y=85.48,
-      size_z=0,  # TODO: measure
-      pedestal_size_z=0,
-      child_location=Coordinate.zero(),  # TODO: measure
-    )
-    self.assign_child_resource(self.plate_holder, location=Coordinate.zero())
-
-  def serialize(self) -> dict:
-    return {**Resource.serialize(self), **Device.serialize(self)}
-
-  async def open(self, slow: bool = False) -> None:
-    await self.driver.open(slow=slow)
-
-  async def close(self, slow: bool = False) -> None:
-    await self.driver.close(slow=slow)
+from .cytation1 import Cytation1 as Cytation1  # noqa: F401
+from .cytation5 import Cytation5 as Cytation5  # noqa: F401
